Fix templates
This pairs with a PR I submitted for the WebApiClient to fix the
template changes that were implemented for .NET core.  Any endpoints
that require a '+' or other reserved symbols in the spec, need to
explicitly allow them in the template.  I believe this is all of the
current ApiEndpoints that are affected